Move the intercept remote mounters into a remotefs package.
Makes the remotefs more generic and also exports them so that they can
be used from other modules.
